# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Universal Curriculum experimental API
  version: 0.2.8
  description: |
    Experimental API for the 0.x release line. This unversioned contract may
    change between minor releases and does not yet provide compatibility
    guarantees. Browser CORS access is not enabled. JSON request bodies are
    limited to 1 MiB.
servers:
  - url: /api
security: []
...
paths:
...
  /curriculum:
    get:
      tags: [Curriculum]
      operationId: getCurriculum
      summary: Get the current published curriculum
      responses:
        '200':
          description: Current curriculum graph and publication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Curriculum'
        '500': {$ref: '#/components/responses/InternalError'}
  /units:
    get:
      tags: [Curriculum]
      operationId: listUnits
      summary: Search and list current units
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: query
          in: query
          schema: {type: string, maxLength: 200}
      responses:
        '200':
          description: Current units
          content:
            application/json:
              schema:
                type: object
                required: [units, page]
                properties:
                  units:
                    type: array
                    items: {$ref: '#/components/schemas/UnitSummary'}
                  page: {$ref: '#/components/schemas/Page'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '500': {$ref: '#/components/responses/InternalError'}
  /units/{unitId}:
    parameters:
      - $ref: '#/components/parameters/UnitID'
    get:
      tags: [Curriculum]
      operationId: getUnit
      summary: Get current unit content and relationships
      responses:
        '200':
          description: Unit detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Unit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /curriculum/proposals:
    get:
      tags: [Curriculum]
      operationId: listAcceptedProposals
      summary: List accepted publications newest first
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Accepted proposal history
          content:
            application/json:
              schema:
                type: object
                required: [proposals, page]
                properties:
                  proposals:
                    type: array
                    items: {$ref: '#/components/schemas/ProposalSummary'}
                  page: {$ref: '#/components/schemas/Page'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '500': {$ref: '#/components/responses/InternalError'}
  /curriculum/proposals/{proposalId}:
    parameters:
      - $ref: '#/components/parameters/ProposalID'
    get:
      tags: [Curriculum]
      operationId: getAcceptedProposal
      summary: Get one accepted publication and its changes
      responses:
        '200':
          description: Accepted proposal
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Proposal'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /learning-paths:
    get:
      tags: [Learning]
      operationId: listLearningPaths
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: The token owner's learning paths
          content:
            application/json:
              schema:
                type: object
                required: [learning_paths]
                properties:
                  learning_paths:
                    type: array
                    items: {$ref: '#/components/schemas/LearningPath'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
    post:
      tags: [Learning]
      operationId: createLearningPath
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LearningPathInput'}
      responses:
        '201':
          description: Created path
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningPath'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
  /learning-paths/{pathId}:
    parameters:
      - $ref: '#/components/parameters/PathID'
    get:
      tags: [Learning]
      operationId: getLearningPath
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Owned learning path
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningPath'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
    put:
      tags: [Learning]
      operationId: replaceLearningPath
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LearningPathInput'}
      responses:
        '200':
          description: Updated path
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningPath'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
    delete:
      tags: [Learning]
      operationId: deleteLearningPath
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Deleted}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /recommendations:
    get:
      tags: [Learning]
      operationId: listRecommendations
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Available next units grouped by learning path
          content:
            application/json:
              schema:
                type: object
                required: [recommendations]
                properties:
                  recommendations:
                    type: array
                    items: {$ref: '#/components/schemas/Recommendation'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
  /progress:
    get:
      tags: [Learning]
      operationId: getProgress
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Direct and recognized completion state
          content:
            application/json:
              schema:
                type: object
                required: [progress]
                properties:
                  progress:
                    type: array
                    items: {$ref: '#/components/schemas/Progress'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
  /progress/{unitId}:
    parameters:
      - $ref: '#/components/parameters/UnitID'
    put:
      tags: [Learning]
      operationId: setProgress
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [completed]
              properties:
                completed: {type: boolean}
      responses:
        '200':
          description: Updated completion
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Progress'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /proposals:
    get:
      tags: [Proposals]
      operationId: listProposals
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          schema: {type: string, enum: [draft, submitted, accepted, rejected]}
      responses:
        '200':
          description: Accepted proposals and private proposals visible to the contributor
          content:
            application/json:
              schema:
                type: object
                required: [proposals, page]
                properties:
                  proposals:
                    type: array
                    items: {$ref: '#/components/schemas/ProposalSummary'}
                  page: {$ref: '#/components/schemas/Page'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
    post:
      tags: [Proposals]
      operationId: createProposal
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProposalInput'}
      responses:
        '201':
          description: Created draft
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Proposal'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
  /proposals/{proposalId}:
    parameters:
      - $ref: '#/components/parameters/ProposalID'
    get:
      tags: [Proposals]
      operationId: getProposal
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Proposal detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Proposal'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
    put:
      tags: [Proposals]
      operationId: replaceProposalMetadata
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProposalInput'}
      responses:
        '200':
          description: Updated draft
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Proposal'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
    delete:
      tags: [Proposals]
      operationId: deleteProposal
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Deleted draft}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /proposals/{proposalId}/units:
    parameters:
      - $ref: '#/components/parameters/ProposalID'
    post:
      tags: [Proposals]
      operationId: createProposalUnit
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, content]
              properties:
                name: {type: string, minLength: 1, maxLength: 200}
                content: {type: string, minLength: 1}
      responses:
        '201':
          description: Proposed unit
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Unit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /proposals/{proposalId}/units/{unitId}:
    parameters:
      - $ref: '#/components/parameters/ProposalID'
      - $ref: '#/components/parameters/UnitID'
    put:
      tags: [Proposals]
      operationId: replaceProposalUnit
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, content]
              properties:
                name: {type: string, minLength: 1, maxLength: 200}
                content: {type: string, minLength: 1}
      responses:
        '200':
          description: Updated proposed unit
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Unit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
    delete:
      tags: [Proposals]
      operationId: deleteProposalUnit
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Proposed deletion recorded}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /proposals/{proposalId}/dependencies:
    parameters:
      - $ref: '#/components/parameters/ProposalID'
    post:
      tags: [Proposals]
      operationId: addProposalDependency
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DependencyInput'}
      responses:
        '204': {description: Dependency addition recorded}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /proposals/{proposalId}/dependencies/{unitId}/{prerequisiteId}:
    parameters:
      - $ref: '#/components/parameters/ProposalID'
      - $ref: '#/components/parameters/UnitID'
      - $ref: '#/components/parameters/PrerequisiteID'
    delete:
      tags: [Proposals]
      operationId: removeProposalDependency
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Dependency removal recorded}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /proposals/{proposalId}/recognitions:
    parameters:
      - $ref: '#/components/parameters/ProposalID'
    post:
      tags: [Proposals]
      operationId: addProposalRecognition
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [source_unit_ids, target_unit_ids]
              properties:
                source_unit_ids:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items: {type: integer, format: int64, minimum: 1}
                target_unit_ids:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items: {type: integer, format: int64, minimum: 1}
      responses:
        '204': {description: Recognition recorded}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /proposals/{proposalId}/changes/{changeId}:
    parameters:
      - $ref: '#/components/parameters/ProposalID'
      - $ref: '#/components/parameters/ChangeID'
    delete:
      tags: [Proposals]
      operationId: deleteProposalChange
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Change deleted}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: uc_api token
  parameters:
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 25}
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    UnitID:
      name: unitId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    Page:
      type: object
      required: [limit, offset, count, has_more]
      properties:
        limit: {type: integer}
        offset: {type: integer}
        count: {type: integer}
        has_more: {type: boolean}
    UnitSummary:
      type: object
      required: [id, name, retired]
      properties:
        id: {type: integer, format: int64}
        name: {type: string, minLength: 1, maxLength: 200}
        retired: {type: boolean}
...
    Curriculum:
      type: object
      required: [proposal_id, units, dependencies]
      properties:
        proposal_id:
          type: [integer, 'null']
          format: int64
        units:
          type: array
          items: {$ref: '#/components/schemas/Unit'}
        dependencies:
          type: array
          items: {$ref: '#/components/schemas/Dependency'}
...